Fix misc MSVC code analysis warnings

+++ src/overmap.cpp
@@ -139,7 +139,7 @@ constexpr size_t set_segment( size_t line, om_direction::type dir )
     if( dir == om_direction::type::invalid ) {
         return line;
     }
-    return line | 1 << static_cast<int>( dir );
+    return line | 1 << static_cast<size_t>( dir );
 }
 
 constexpr bool has_segment( size_t line, om_direction::type dir )
@@ -147,7 +147,7 @@ constexpr bool has_segment( size_t line, om_direction::type dir )
     if( dir == om_direction::type::invalid ) {
         return false;
     }
-    return static_cast<bool>( line & 1 << static_cast<int>( dir ) );
+    return static_cast<bool>( line & 1 << static_cast<size_t>( dir ) );
 }
 
 constexpr bool is_straight( size_t line )
